share plugins via project base
