+++ packages/runner-ct/lib/runner-ct.ts
@@ -2,7 +2,8 @@ import Debug from 'debug'
 import _ from 'lodash'
 import path from 'path'
 import send from 'send'
-import { Cfg, ProjectCt, SpecsStore } from '@packages/server-ct'
+import { Cfg, ProjectCt } from '@packages/server-ct'
+import { SpecsStore } from '@packages/server/lib/specs-store'
 
 interface ServeOptions {
   config: Cfg
