+++ packages/server-ct/src/socket-ct.ts
@@ -33,8 +33,4 @@ export class SocketCt extends SocketBase {
       },
     })
   }
-
-  sendSpecList (specs) {
-    this.toRunner('component:specs:changed', specs)
-  }
 }
