+++ packages/server/lib/project-e2e.ts
@@ -1,11 +1,9 @@
 import Debug from 'debug'
 import browsers from './browsers'
 import config from './config'
-import plugins from './plugins'
 import preprocessor from './plugins/preprocessor'
 import { ProjectBase } from './project-base'
 import { ServerE2E } from './server-e2e'
-import settings from './util/settings'
 
 const debug = Debug('cypress:server:project')
 
@@ -49,26 +47,16 @@ export class ProjectE2E extends ProjectBase<ServerE2E> {
     })
   }
 
+  _onError<Options extends Record<string, any>> (err: Error, options: Options) {
+    debug('got plugins error', err.stack)
+
+    browsers.close()
+
+    options.onError(err)
+  }
+
   _initPlugins (cfg, options) {
-    // only init plugins with the
-    // allowed config values to
-    // prevent tampering with the
-    // internals and breaking cypress
-    cfg = config.allowed(cfg)
-
-    return plugins.init(cfg, {
-      projectRoot: this.projectRoot,
-      configFile: settings.pathToConfigFile(this.projectRoot, options),
-      testingType: options.testingType,
-      onError (err) {
-        debug('got plugins error', err.stack)
-
-        browsers.close()
-
-        options.onError(err)
-      },
-      onWarning: options.onWarning,
-    })
+    return super._initPlugins(cfg, options)
   }
 
   close () {
